+++ integration/spark/spark3/src/main/java/io/openlineage/spark3/agent/lifecycle/plan/DataSourceV2ScanRelationInputDatasetBuilder.java
@@ -9,8 +9,8 @@ import io.openlineage.client.OpenLineage;
 import io.openlineage.spark.api.AbstractQueryPlanInputDatasetBuilder;
 import io.openlineage.spark.api.DatasetFactory;
 import io.openlineage.spark.api.OpenLineageContext;
+import io.openlineage.spark3.agent.utils.DataSourceV2RelationDatasetExtractor;
 import io.openlineage.spark3.agent.utils.DatasetVersionDatasetFacetUtils;
-import io.openlineage.spark3.agent.utils.PlanUtils3;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan;
@@ -41,6 +41,7 @@ public class DataSourceV2ScanRelationInputDatasetBuilder
         context.getOpenLineage().newDatasetFacetsBuilder();
 
     DatasetVersionDatasetFacetUtils.includeDatasetVersion(context, datasetFacetsBuilder, relation);
-    return PlanUtils3.fromDataSourceV2Relation(factory, context, relation, datasetFacetsBuilder);
+    return DataSourceV2RelationDatasetExtractor.extract(
+        factory, context, relation, datasetFacetsBuilder);
   }
 }
