+++ integration/spark/spark33/src/main/java/io/openlineage/spark33/agent/lifecycle/plan/ReplaceIcebergDataDatasetBuilder.java
@@ -8,8 +8,8 @@ package io.openlineage.spark33.agent.lifecycle.plan;
 import io.openlineage.client.OpenLineage;
 import io.openlineage.spark.api.AbstractQueryPlanOutputDatasetBuilder;
 import io.openlineage.spark.api.OpenLineageContext;
+import io.openlineage.spark3.agent.utils.DataSourceV2RelationDatasetExtractor;
 import io.openlineage.spark3.agent.utils.DatasetVersionDatasetFacetUtils;
-import io.openlineage.spark3.agent.utils.PlanUtils3;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -66,7 +66,7 @@ public class ReplaceIcebergDataDatasetBuilder
       DatasetVersionDatasetFacetUtils.includeDatasetVersion(context, datasetFacetsBuilder, table);
     }
 
-    return PlanUtils3.fromDataSourceV2Relation(
+    return DataSourceV2RelationDatasetExtractor.extract(
         outputDataset(), context, table, datasetFacetsBuilder);
   }
